vector-demo/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.29
psycopg2-binary==2.9.9
pgvector==0.2.4
pymilvus==2.4.4
python-dotenv==1.0.1
